Add files via upload
Random Name Picker
One of my main responsibilities as project manager was to coordinate the process of deliverables and I had to randomly assign reviewing tasks to contributors. This task inspired the program "Random Name Picker" 

This python program allows the user to randomly select a specified number of names from a text file. The names in the file should be separated by semicolons ";". After reading the file, the program prompts the user to enter how many names they want to randomly choose, and then displays the selected names.